refactor: pull out verification code input
